pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_backend"

[project]
name = "stonkers"
version = "0.1.0"
description = "Modular algorithmic trading bot for cryptocurrency markets"
requires-python = ">=3.11"
dependencies = [
    "ccxt>=4.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "ta>=0.10.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
